+++ cli/CHANGELOG.md
@@ -3,6 +3,9 @@
 
 _Released 05/23/2023 (PENDING)_
 
+**Features:**
+- Add Git related messages for the Runs and Debug Page when users aren't using Git or their are no recorded runs for the current branch. Fixes [#26680](https://github.com/cypress-io/cypress/issues/26680).
+
 **Bugfixes:**
 
 - Reverted [#26452](https://github.com/cypress-io/cypress/pull/26630) which introduced a bug that prevents users from using End to End with Yarn 3. Fixed in [#26735](https://github.com/cypress-io/cypress/pull/26735). Fixes [#26676](https://github.com/cypress-io/cypress/issues/26676).