+++ cli/CHANGELOG.md
@@ -4,6 +4,7 @@
 _Released 05/23/2023 (PENDING)_
 
 **Features:**
+
 - Add Git related messages for the Runs and Debug Page when users aren't using Git or their are no recorded runs for the current branch. Fixes [#26680](https://github.com/cypress-io/cypress/issues/26680).
 
 **Bugfixes:**