+++ cli/CHANGELOG.md
@@ -9,6 +9,7 @@ _Released 05/23/2023 (PENDING)_
 
 **Bugfixes:**
 
+- Fixed an issue where children of HTML elements with `display: contents; overflow: hidden` were not visible by Cypress tests. Fixes [#25199](https://github.com/cypress-io/cypress/issues/25199). 
 - Reverted [#26452](https://github.com/cypress-io/cypress/pull/26630) which introduced a bug that prevents users from using End to End with Yarn 3. Fixed in [#26735](https://github.com/cypress-io/cypress/pull/26735). Fixes [#26676](https://github.com/cypress-io/cypress/issues/26676).
 - Moved `types` condition to the front of `package.json#exports` since keys there are meant to be order-sensitive. Fixed in [#26630](https://github.com/cypress-io/cypress/pull/26630).
 - Fixed an issue where newly-installed dependencies would not be detected during Component Testing setup. Addresses [#26685](https://github.com/cypress-io/cypress/issues/26685).